backend/pyproject.toml
[project]
name = "dnd-scheduler"
version = "0.1.0"
description = "DnD session scheduler backend"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13",
    "bcrypt>=4.1",
    "email-validator>=2.3",
    "fastapi>=0.110",
    "google-auth>=2.29",
    "httpx>=0.27",
    "passlib>=1.7",
    "psycopg[binary]>=3.1",
    "pydantic>=2.12",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.21",
    "requests>=2.32",
    "resend",
    "sqlalchemy>=2.0",
    "uvicorn>=0.29",
    "watchfiles>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
]

[tool.uv]
# Не устанавливать backend как пакет — это приложение, а не библиотека
package = false
